Move `rustdoc::clean::utils::find_nearest_parent_module` to `TyCtxt`
`find_nearest_parent_module` was extracted from
`rustdoc::passes::collect_intra_doc_links` in #80368.

We (me and Joshua Nelson) thought at the time that it should be in rustc
instead, and Joshua suggested it be a method on `TyCtxt`.
However, since #80368 was fixing a significant regression that was about
to land on beta, we agreed that to be able to fix this quickly I should
leave the code in rustdoc.

This is a followup PR to move it to `TyCtxt`.