Deprecate std::path::Path's aliases of `std::fs` functions.
`std::path::Path` has several functions which are trivial aliases for `std::fs`
functions (eg. [`metadata`]) and a few which are minor convenience wrappers for
`std::fs` functions (eg. [`is_file`]).

[`metadata`]: https://doc.rust-lang.org/std/path/struct.Path.html#method.metadata
[`is_file`]: https://doc.rust-lang.org/std/path/struct.Path.html#method.is_file

This PR deprecates these functions, so that `std` only has one preferred name
for these operations. And it clarifies the role of the `std::path` module, which
otherwise just has pure data structures and algorithms.

The convenience wrappers, `is_file`, `is_dir`, and `exists`, are operations for
which seemingly obvious uses are prone to TOCTTOU errors. So while it would be
straightforward to add corresponding similar convenience functions to `std::fs`,
this PR does not propose doing so.

Specifically, this PR deprecates the 5 functions in `std::path::Path` which are
trivial aliases of functions in `std::fs`:
 - [`metadata`](https://doc.rust-lang.org/std/path/struct.Path.html#method.metadata)
 - [`symlink_metadata`](https://doc.rust-lang.org/std/path/struct.Path.html#method.symlink_metadata)
 - [`canonicalize`](https://doc.rust-lang.org/std/path/struct.Path.html#method.canonicalize)
 - [`read_link`](https://doc.rust-lang.org/std/path/struct.Path.html#method.read_link)
 - [`read_dir`](https://doc.rust-lang.org/std/path/struct.Path.html#method.read_dir)

and the 3 which are convenience wrappers around `std::fs::metadata`:
 - [`is_file`](https://doc.rust-lang.org/std/path/struct.Path.html#method.is_file)
 - [`is_dir`](https://doc.rust-lang.org/std/path/struct.Path.html#method.is_dir)
 - [`exists`](https://doc.rust-lang.org/std/path/struct.Path.html#method.exists)