Refactor final_arg_types to use a presized array
Previously, it was a tuple including the index of the argument it was
discussing, but this made it awkward to look up the types later.  The
primary purpose of this was to pass it to another helper function, so I
refactored that as well.

Rather than trying to figure out the iterator chain that would work with
the new structure, I just wrote the loop manually.