Fixed a bug related to invalid arguments
Consider this case:

```
fn some_func(a: i32, b: i32) {}
some_func(1, "");
```

Before this change, the first 1 would register as able to satisfy
the second parameter, so the "" would show up as "extra" and we'd remove
it.

We would then allow the first argument to satisfy the first parameter,
and the second parameter would have no argument, so it would show up as
"missing".

A simpler way to express this is just that the type of the second
argument is invalid.  By doing one pass of satisfying arguments first,
we can handle this case.