A fix for a &str type check, and a pass at extra span alignment
There were some test cases where "" wasn't typechecking against &str in
the "soft" typechecking path (the one that doesn't persist new
constraints.)

I found a method that appears to typecheck without persisting new
constraints, but I can't be sure, so I'll need to double check that.
For now, this new expression typechecking method works well.

This also tries to carefully handle eliminating extra arguments.  Spans
are tricky, as we need to eliminate commas correctly, and the suggestion
code gets angry if we have overlapping spans, so this tries to find
contiguous blocks to eliminate.

I'll probably simplify this now that I understand the logic a bit
better, but it's late, so I'm going to bed.