Redirect to "main crate" when clicking Rust logo (or custom logo)
The "main crate" in this case would be the one that is opened by `cargo
doc --open` by default. However, since doing this on cargo's end would
require a lot more work, @jyn514 recommended a hacky solution that
works. This solution is to create a redirect page (called
`logo-redirect.html`) in the documentation root directory when
documenting a crate. Thus, the last crate documented (usually a binary
or the main library) is the one redirected to by this file. This file is
then pointed to by the logo.

Note that as jynelson noted:
> this assumes you never document examples, and that if you
> document a binary, you want it to take precedence over the library

I'm pretty sure this is probably the desired behaviour (and I don't
think that examples are documented during `cargo doc`), so this hack
should probably work for now. (And that's how programming works!)