Add a parse recovery in array type syntax
When user types [x, y] in a type context instead of [x; y] we now try to
recover. In particular this fixes confusing "unmatched angle bracket"
errors in code like

    fn main() {
        drop::<[(), 0]>([]);
    }

which would previously generate

     --> src/main.rs:2:11
      |
    2 |     drop::<[(), 0]>([]);
      |           ^ help: remove extra angle bracket

    error: expected one of `;` or `]`, found `,`
     --> src/main.rs:2:15
      |
    2 |     drop::<[(), 0]>([]);
      |               ^ expected one of `;` or `]`

With the recovery we now only generate the second error.

Fixes #81097

Note that we only try this recovery in edition 2018 and later. In
earlier editions functions without bodies are now allowed to have named
arguments, so having this recovery is causing problems in cases like

    trait Foo {
        fn foo([a, b]: [i32; 2]) {}
    }

where the `[a, b]` part is an error, and assuming `;` in place of `,`
there makes the error message worse, not better. A test for this is in
`ui/issues/issue-50571.rs`.