Remove `Iterator::is_sorted_by_key`
This method can be trivially replaced by `iter.map(...).is_sorted()`.
There has been some discussion about this, but the only reason for
having that method that has been brought up is to have a more symmetry
between the methods of `[T]` and `Iterator`. `[T]::is_sorted_by_key` is
necessary, as there is not a simple replacement for it. But I don't
think API symmetry is a strong enough argument for adding methods that
are essentially superfluous or "a tiny bit convenient" at most.

Finally, this method can always be added in the future. For the initial
stabilization of the `is_sorted` API, it is not necessary.