Change `is_sorted[_by_key]` methods to require `Ord` instead of `PartialOrd`
With `f32::total_cmp` and `f64::total_cmp`, the main reason for having
an `PartialOrd` bound on these methods is mostly gone. By requiring
`Ord`, we have no more edges cases with non-comparable values that need
to be explained in the docs. These methods are now also in line with
the `sort*` methods.

Also note that `is_sorted` is mostly for convenience; if it is not
powerful enough, there are more powerful tools that are only slightly
less convenient. Finally, the trait bound can be relaxed in the future.