Add back error span labels
@estebank suggested we only annotate these if there are 5 or less.

In addition to the suggestion at the bottom of the error, we label the
specific problems we encounter without suggesting how to fix it.

This also ladders the suggestion text: if there's just one thing to
suggest, use specific language about the suggestion; however, if there
are multiple suggestions, use a generic "did you mean"

Even though the comment on span_suggestion suggests against this,
Camelid on zulip pointed out that there are already a few suggestions
like this, and any other language sounds slightly awkward.