Change debug assertions to actual asserions
I think incorrect (or malicious) `PartialEq` and `PartialOrd` implementations for `T` could define `x == y` and `x < y` to both be conditionally true with interior mutability in such a way that `pos == len - 1` (implying `data[pos] == data[len - 1]`) but  we end up in the `data[pos] < data[len - 1]` branch.  `sift_up()` and `sift_down()` assume that `pos` is in-bounds which would be violated in this case.