Stop leaking implementation detail in `Debug` impls of opaque types
This changes the `Debug` impl of:
- `core::any::TypeId`
- `core::mem::Discriminant`
- `std::thread::ThreadId`
- `std::time::Instant`

All these types are explicitly described as "opaque" in the
documentation, so they shouldn't print implementation details in their
`Debug` implementation.

The change for `ThreadId` can possibly be reverted in the future, once
the method `thread_id_value` is stabilized:
https://github.com/rust-lang/rust/issues/67939
The same goes for `Discriminant` once/if the `DiscriminantKind` trait
gets stabilized.
But until then, these are opaque types.