Inline the `core::arch` documentation into `std`
This commit removes the `#[doc(no_inline)]`-attribute from the `pub use`
statement of the `core::arch`-module. This has the effect, that the
module is not listed under the re-exported section in the documentation
(it is the only item in a re-exports-section in the `std`), but as a
normal module.
This also removes the only mention of the `core`-crate in the `std`
crate, which is desirable, as only a subset of Rust users know `core`
and its relationship with the `std`.