Sort bounds by DefPathHash during astconv
cc #82920

The value of a `DefId` is not stable across compilation sessions. If we
sort by `DefIds`, and then include the sorted list as part of a query
result, then the query will no longer be a pure function of its input
(e.g. the `DefId` can change while the `DefPathHash` remains the same).

For reasons that are not yet clear, this issue lead to segfaults and
garbage slice index values when the project in the linked issue was
built with a particular incremental cache.