Unused var anlysis: don't consider uninhabited return value as early return
Previously calling a function that returns in uninhabited type (`!` or
`Infallible`) would be considered as an early return. This causes
"unused variable" warnings when the binding for the return value is
actually used:

    fn never() -> core::convert::Infallible {
        panic!()
    }

    fn main() {
        let n = never();
        match n {} // generates "unused variable" warning
    }

It doesn't make sense to generate an "unused variable" warning for
used variables, so we now don't consider uninhabited returns as early
return and avoid generating a warning in these cases.

- Existing test 'never-assign-dead-code' updated
- Regression test for #83276 added

Fixes #83276