Implement bool.take()
I found myself wanting to write code using this method on `bool`. It essentially mirrors `Option::<()>::take`

This would make it easy to both check and reset a `bool` value in one line, ex `if need_stuff.take() { // do stuff }` 

If this seems reasonable, I will open a tracking issue. If not, feel free to just close this.