Add intra-doc `field@` disambiguator
This change required some refactoring because fields (1) don't get their
own `Res` (instead they seem to use the `Res` of their parent node) and
(2) `DefKind::Field::ns()` returns `None`, so rustdoc has to
special-case fields. For some reason, we treat fields as if they're in
the value namespace - I'm not sure why that is.