Add `core::array::from_fn`
This adds `core::array::from_fn` for initializing an array from a
closure. The API is meant to closely resemble `core::iter::from_fn` with
the difference that the array dictates the amount of elements produced
as opposed to the function.

This should also cover a lot of cases where one might want to reach for
collecting into an array, but without all the associated problems that
this would have, such as providing too few or too many elements.

```rust
let array = std::array::from_fn(|index| 2 * index);
assert_eq!(array, [0, 2, 4, 6]);
```