Add assertion details to panic info.
The assertion details are added as a new enum field, allowing for future
extensions with different information for other sources of panics.
For example, `.unwrap()` could extend the enum to store a `&dyn Error`
in the panic info.

The panic handler in `std` passes the additional info unmodified to the
panic hook.