Use `EvaluatedToOkModuloRegions` when we get a coinductive match
This forces every participant in the cycle to have a result
`EvaluatedToOkModuloRegions` (or an error) as well, since their result
will be influenced by the coinductive result.

While I'm confident that this is sound (AFAIK, it's always sound to
replace `EvaluatedToOk` with `EvaluatedToOkModuloRegions`), I don't
really understand the coinductive cycle logic. This appears to fix the
`syn` incremental crash, but there could be a different way of fixing it
as well.