std: Fix st_lspare issue with FreeBSD 13 ABI
The check for whether or not st_lspare exists was only checking against
freebsd12, not freebsd12 or any newer.

Since freebsd11 is the oldest supported, flip the sense of the check
and use the old implemention for freebsd11 instead of not(freebsd12).

This fixes building against freebsd13 ABI and future-proofs the check.