rustdoc: add support for codeblock_attr param for doc attributes
The `codeblock_attr` parameter for `#[doc]` overrides the codeblock
attributes - that is, the text `foo` after ```` ```foo ````.

This also allows a codeblock attribute to be applied to indentation
code blocks.

This is intended for documentation coming from non-Rust-aware sources
which may contain code examples, but are not written in Rust. Such
examples - at best - will cause compilation warnings from rustdoc,
or could compile but have unintended consequences.