Fix positioning for detail toggles on mobile.
The rule that was supposed to put these in the right place on mobile had
an extra class (item-list) in its selector, but that class isn't
actually part of the DOM for everything we care about.

After removing that extra class, I found the rule wasn't firing. That's
because @media doesn't increase specificity
(https://css-tricks.com/how-much-specificity-do-rules-have-like-keyframes-and-media/),
and the rule was being overridden by the later rule with the exact same
selector.

I moved all the details-related rules above the @media queries because
in general I think we expect rules inside @media to override others.
However, I didn't move the rules that I haven't touched, since I don't
know all the places to check that they are working as expected.