Add `Option::merge` under `option_merge` feature gate
This patch adds unstable `Option::merge` method which takes two `Option<T>`s
and a `T -> T -> T` function. If only one of the options is `Some(_)` the
method returns it. If both options are `None` the method returns `None`.
Otherwise, if both options are `Some(_)` the method merges them by applying
the given function.