Fix `x.py doc --stage 1 src/tools/error_index_generator`
The error index is a complicated tool (because it depends on rustdoc)
and this fix is only one of many possible approaches.

Here is the sequence of fixes that culminated in this commit:

1. The error index gives an error that libLLVM-nightly.so isn't found.
2. libLLVM-nightly.so is present in stage1/lib, but not in stage0/lib.
3. `builder.sysroot()` returns `stage0-sysroot`, but
   `builder.rustc_libdir()` returns `stage0/lib`.
4. Therefore, special case the sysroot to be `stage0` for the error index.

Another, possibly better fix is to stop depending on rustdoc at all, and
call it as a binary or separate out a shared crate. But that's a larger
refactor.