read_target_uint => read_target_ptr (nfc)
Most public callers of read_target_uint are using it for the purpose
of acquiring a pointer-sized uint. Using read_target_ptr instead
makes this usage more clear, and exploits the knowledge of the caller
to make the function into a clear "table lookup".

This enables making read_target_uint private, and is intended to help
simplify refactoring the interpreter in rustc_middle.

In cg_clif, one usage of read_target_uint was to obtain a u32, which
has been specialized to match the exact numeric width.